+++ setup.cfg
@@ -33,7 +33,7 @@ install_requires =
 	inspyred
 	jmetalpy
 	networkx
-	matplotlib
+	matplotlib<=3.5.0
 	tqdm
 	joblib
 	httpx~=0.23
